requirements.txt
python-telegram-bot
pyTelegramBotAPI
psycopg2
argparse
requests==2.7.0
configparser
bs4
selenium
pytest-shutil
logger
lxml
boto